requirements.txt
BeautifulSoup==3.2.1
Flask==0.10.1
Flask-Cache==0.12
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
distribute==0.6.24
gunicorn==18.0
ipdb==0.8
itsdangerous==0.23
lxml==3.2.3
python-cjson==1.0.5
requests==2.0.0
wsgiref==0.1.2
